chore: rename src/external to src/languageSupport
