+++ jestSetup.ts
@@ -32,7 +32,7 @@ jest.mock('src/shared/utils/errors')
   the mocked values in src/external/languages/flux/parser are ignored.
   So, need to mock here as well
 */
-jest.mock('src/external/languages/flux/parser', () => ({
+jest.mock('src/languageSupport/languages/flux/parser', () => ({
   parse: jest.fn(getMockedParse()),
   format_from_js_file: jest.fn(),
 }))
