+++ src/flows/pipes/Notification/NotificationMonacoEditor.tsx
@@ -5,8 +5,8 @@ import React, {FC} from 'react'
 import MonacoEditor from 'react-monaco-editor'
 
 // Utils
-import LANGID from 'src/external/languages/markdown/monaco.markdown.syntax'
-import THEME_NAME from 'src/external/languages/flux/monaco.flux.theme'
+import LANGID from 'src/languageSupport/languages/markdown/monaco.markdown.syntax'
+import THEME_NAME from 'src/languageSupport/languages/flux/monaco.flux.theme'
 
 // Types
 import {EditorType} from 'src/types'
