+++ src/languageSupport/languages/flux/keybindings.test.ts
@@ -1,7 +1,7 @@
 import {
   isCommented,
   toggleCommenting,
-} from 'src/external/languages/flux/monaco.flux.hotkeys'
+} from 'src/languageSupport/languages/flux/monaco.flux.hotkeys'
 
 const TEST_CASES = [
   ['//moo', true],
