+++ src/languageSupport/languages/flux/monaco.flux.hotkeys.ts
