+++ src/languageSupport/languages/markdown/monaco.markdown.syntax.ts
@@ -1,11 +1,11 @@
-import register from 'src/external/monaco.onigasm'
+import register from 'src/languageSupport/monaco.onigasm'
 
 const LANGID = 'markdown'
 
 register(LANGID, async () => ({
   format: 'json',
   content: await import(
-    /* webpackPrefetch: 0 */ 'src/external/languages/markdown/markdown.tmLanguage.json'
+    /* webpackPrefetch: 0 */ 'src/languageSupport/languages/markdown/markdown.tmLanguage.json'
   ).then(data => {
     return JSON.stringify(data)
   }),
