+++ src/languageSupport/languages/python/monaco.python.syntax.ts
@@ -1,11 +1,11 @@
-import register from 'src/external/monaco.onigasm'
+import register from 'src/languageSupport/monaco.onigasm'
 
 const LANGID = 'python'
 
 register(LANGID, async () => ({
   format: 'json',
   content: await import(
-    /* webpackPrefetch: 0 */ 'src/external/languages/python/python.tmLanguage.json'
+    /* webpackPrefetch: 0 */ 'src/languageSupport/languages/python/python.tmLanguage.json'
   ).then(data => {
     return JSON.stringify(data)
   }),
