+++ src/languageSupport/languages/toml/monaco.toml.syntax.ts
@@ -1,11 +1,11 @@
-import register from 'src/external/monaco.onigasm'
+import register from 'src/languageSupport/monaco.onigasm'
 
 const LANGID = 'toml'
 
 register(LANGID, async () => ({
   format: 'json',
   content: await import(
-    /* webpackPrefetch: 0 */ 'src/external/languages/toml/toml.tmLanguage.json'
+    /* webpackPrefetch: 0 */ 'src/languageSupport/languages/toml/toml.tmLanguage.json'
   ).then(data => {
     return JSON.stringify(data)
   }),
