+++ src/shared/components/TomlMonacoEditor.tsx
@@ -3,8 +3,8 @@ import React, {FC} from 'react'
 
 // Components
 import MonacoEditor from 'react-monaco-editor'
-import THEME_NAME from 'src/external/languages/toml/monaco.toml.theme'
-import TOMLLANGID from 'src/external/languages/toml/monaco.toml.syntax'
+import THEME_NAME from 'src/languageSupport/languages/toml/monaco.toml.theme'
+import TOMLLANGID from 'src/languageSupport/languages/toml/monaco.toml.syntax'
 import {OnChangeScript} from 'src/types/flux'
 import * as monacoEditor from 'monaco-editor/esm/vs/editor/editor.api'
