+++ src/tasks/actions/thunks.ts
@@ -30,7 +30,10 @@ import {
 
 // Constants
 import * as copy from 'src/shared/copy/notifications'
-import {parse, format_from_js_file} from 'src/external/languages/flux/parser'
+import {
+  parse,
+  format_from_js_file,
+} from 'src/languageSupport/languages/flux/parser'
 
 // Types
 import {
